+++ t/check_sas2ircu.t
@@ -6,15 +6,27 @@ BEGIN {
 
 use strict;
 use warnings;
-use Test::More tests => 5;
+use Test::More tests => 10;
 use test;
 
 my @tests = (
+##removed for the moment because I dont have an env w/ aprops display to match list and statu
+	## no volumes, 12 drives in ready state
+	{
+		status => OK,
+		list => 'pr71/sas2ircu-list.out',
+		cstatus => 'pr71/sas2ircu-status.out',
+		cdisplay => 'pr71/sas2ircu-display.out',
+		message => 'ctrl #0: 0 Vols: No Volumes: 12 Drives: Ready (RDY)::',
+	},
 	{
 		status => OK,
 		list => 'test1/LIST.log',
 		cstatus => 'test1/0-STATUS.log',
-		message => 'ctrl #0: Optimal',
+## this file is just a mockup based on the above test since I dont have original output for this case
+		cdisplay => 'test1/sas2ircu-display.out',
+#and similarly mocked up
+	 	message => 'ctrl #0: 1 Vols: Optimal: 3 Drives: Ready (RDY)::',
 	},
 );
 
@@ -24,6 +36,7 @@ foreach my $test (@tests) {
 		commands => {
 			'controller list' => ['<', TESTDIR . '/data/sas2ircu/' . $test->{list} ],
 			'controller status' => ['<', TESTDIR . '/data/sas2ircu/' . $test->{cstatus} ],
+			'device status' => ['<', TESTDIR . '/data/sas2ircu/' . $test->{cdisplay} ],
 		},
 	);
 	ok($plugin, "plugin created");